Allow alternate CentOS and EPEL mirrors in ISO creation
The yum repositories set for ISO creation in config.yaml are passed to
Pungi and used to download the RPM packages through yum. With the main
repositories, the execution of this scripts in our systems were failing
most of the time. Using an alternate (closer) mirror improves the
reliability of the script and the download times.