Alter trigger jobs to report triggered build URL in GitHub pull-request
Currently the jobs that use the GitHub pull request builder plugin
from Jenkins include the trigger build URL in their "vote". That
displays as a link on the pull request commit status.

The trigger jobs are quite "empty" and the information we're really
interested in lies on the jobs they trigger so this patch changes the
URL to be that of the triggered job.

This change also allows other jobs to read the GitHub pull-request
commit status and get the URL for the triggered job to be used in a
pipeline.