Turn uploads into build steps instead of post build steps
This requires archiving the build results with a custom script, because
the Jenkins archive plugin can only be executed as a post build step.
Artifacts that are not uploaded (e.g. logs) can still be kept as a post
build step.

This allows us to wait for the upload to finish and mark the build as
unstable if the upload fails.
It avoids creating symbolic links prematurely in GSA for weekly and
nightly builds, which could allow users to access incomplete
repositories.