Fix github-api-wrapper role
Ansible scripts can be run remotely so using chdir doesn't work
because it runs on the host running the playbook.

This patch fix this issue but limits the ability of moving the
playbook files in (or out of) the infrastructure repo. However that
shouldn't be a problem as moving files is a modification to the git
repo so another could be made to fix the path in the role.