Replace Jenkins copyArtifacts plugin in trigger weekly with a script
Work is in progress to make our packages and ISO builds triggered only
via GitHub pull requests. Because of that, the weekly trigger will
have to use the GitHub API to discover the triggered build id of the
build_host_os to get the build timestamp.

The copyArtifacts plugin cannot use variables defined in another build
step so I'm replacing it with a script.