Fix fetching of pull requests in build job
The pull requests were being fetched from the passed versions repository
URL in the job script and added to the "origin" remote, whose URL was
left unmodified. When the host_os.py script was called, it identified
that the "origin" remote URL did not match the URL passed in the command
line and substituted the "origin" remote, losing all references fetched
from the custom fetch command.

By re-creating the "origin" remote with the correct URL before executing
the custom fetch, the host_os.py script will leave the remote unmodified
and will be able to find the pull requests. The issue happened in the
build that followed a nightly, which modified the versions repository
remote to the bot GitHub fork.