Checkout the build job's commit id in the weekly job script
The build job is triggered asynchronously by the GitHub pull request
that is created by the weekly job. From the time the weekly job starts
to the time the build job starts there might have been updates to the
builds repo branch so we need to checkout the right commit id after
the build is finished to use in the rest of the weekly job's
operations (release notes and git tags creation).