Ensure release notes creation happens only after the build job has finished
Currently the release notes creation is running as soon as the GitHub
pull request that triggers the build job is created, which means that
it is probably being run before the build job has started. If the
builds repo branch moves in the meantime, the release notes will have
the wrong commit id listed.

This patch moves the release notes creation to after the pull request
has been closed, in the majority of times the build will be finished
by then (unless we merge the PR prematurely).