Add the remote name to *_REPO_REFERENCE in trigger jobs
These variables come in pairs, one for builds and one for
versions. When one contains the git reference of a pull request, the
other will contain the branch reference (usually master) of the other
repo. The pull request reference contains the remote name (origin) so
let's make this uniform.

This also helps disambiguate the reference name in case we have
multiple remotes.