Fix versions repository reference passed to build-release-notes command
This was broken since cea30f93947610b5a817ccea76aebed661fe0647, but did
not show up because the versions repository was shared among
"update-versions" and "create-release-notes" and so the command
"git checkout weekly-<timestamp>" had no problems in finding the remote
reference.

Since we now wait to merge the versions pull request before creating the
release notes, it is safer to use the commit ID instead of the branch
name. GitHub makes it easy to delete a branch when merging a pull
request manually and so the branch reference may not exist anymore.