Pass the actual commit hash when triggering build from versions repo
The "sha1" variable contains a ref spec in the form "origin/pr/*/merge"
or "origin/pr/*/head". If the former is used for a weekly build, the
actual commit hash that was used to build the packages will diverge from
the one used to create the release notes and will not be present in the
master branch, since a new merge commit will be created when the PR is
merged.

By changing to "ghprbActualCommit", we enforce that the commit used to
build is consistent with the release notes and present in the master
branch.