Make periodic builds trigger automatically on setup
The trigger must be set when the job is created from
periodic_pipelines.groovy. When the pipeline is executed, the trigger
is overwritten by the properties defined in
pipeline/daily/stages.groovy.

To avoid duplicating the cron expression in those two places, it is set
by Ansible in the constants file and in the seed job.