Create credentials job via Ansible
The seed job might need credentials to access GitHub if an enterprise
instance is used. Executing the seed job before creating the credentials
was also problematic because the pipelines would start monitoring the
repositories, but not have write access to update pull requests statuses,
for example.

By creating the credentials before anything else is executed on Jenkins,
we fix those problems and catch authentication issues earlier in the
setup.