Create seed job by copying the XML file
Most of the Jenkins configurations were already performed by copying the
XML files to the target system. Creating the seed job though the HTTP
API is not so reliable because Jenkins need to be up at the time the
task executes, which requires some additional roles to be added.
Furthermore, Jenkins will most likely need to be restarted because of
other tasks, so the benefit of not having to restart it for this
specific task is usually irrelevant.