Use 'rpm_macros' option when building packages
Defining RPM macros with the 'mock_args' option makes calls to different
RPM commands inconsistent, because some of them do are not executed via
mock. This caused inconsistencies in the 'release' field in the
resulting 'packages.json' files, which is an input to the release page
cration. Since the 'rpm_macros' option expects a dict, it can not be
used in the command line, so we have to pass a modified YAML config file
to the script.