Disable intermediate workspace cleanups in daily builds
The daily builds execute all steps in a single node, so it is useless to
clean the workspace and checkout the repositories multiple times.

The next commit will also modify the build config file, which will be
shared between steps. To preserve it, we need to disable the cleanup.

An argument has been added to the functions buildPackages, buildIso and
uploadArtifacts to disable the cleanup in daily builds, while keeping
the previous behaviour for pull request builds.