Avoid conflicts when pushing changes to github.io repository
When devel and release pipeline are executing simultaneously, since they
reference the same branch in the github.io repository, they may fetch
that branch at the same time and get the same commit. Later, when they
try to push to this branch, the second one was failing because the tip
of the branch had been updated. Re-fetching and rebasing makes a failure
very unlikely, since those are very fast operations. Conflicts on rebase
should never happen because the pipelines create differently named
files.

The same cannot be done for the versions repository, because it would
change the commit, which was already tested. It is also unnecessary
because the pipelines reference different branches.